app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
pyjwt
requests
bs4
selenium